Win64 warnings fixed
git-svn-id: https://svn.apache.org/repos/asf/harmony/enhanced/drlvm/trunk@540841 13f79535-47bb-0310-9956-ffa450edef68

+++ vm/jitrino/src/optimizer/FastArrayFilling.cpp
@@ -111,7 +111,7 @@ FastArrayFillPass::_run(IRManager& irManager)
         }
         //compute number of nodes of the loop
         Nodes loopNodes = info->getLoopNode(node,false)->getNodesInLoop();
-        unsigned sz = loopNodes.size();
+        size_t sz = loopNodes.size();
         if (sz!=3) {
             continue;
         }
